… docs/querying/functions.md
… promql/bench_test.go
… promql/info_test.go
… promql/parser/parse_test.go