**Port of #14495 from v2.55. Agreed w/ @jesusvazquez that I port the feature to main through a cherry-pick PR.**

This _experimental_ MVP of the [PromQL `info` function](https://github.com/prometheus/proposals/pull/37) simplifies the implementation by assuming:

* Only support for the `target_info` metric
* That `target_info`'s identifying labels are `job` and `instance`

`info` is implemented following the same pattern as traditional join queries, and achieves even (significantly) better performance by only selecting RHS (info) series w/ the correct identifying labels (`instance` and `job`). I believe it's possible though to optimize corresponding join queries in the same way, provided that the RHS is a vector selector (the same heuristic should be applicable). The main motivation for following the join query pattern, i.e. perform a second select for corresponding info series and then join, is that info series may be sharded independently of the corresponding non-info series.

Another significant benefit over traditional join queries is that `info` will resolve conflicts between info series with the same identifying labels, that align to the same timestamp, whereas traditional joins will raise an error for this case. As an example, let's say we have the two following `target_info` series, both identified by the `instance` and `job` labels:

1. `{__name__="target_info", instance="1", job="a", pid="1"}`
1. `{__name__="target_info", instance="1", job="a", pid="2"}`

If series 1's last sample is at timestamp 1, and series 2's last sample is at timestamp 2, `info` will need to choose between the two of them for timestamp 2 (since also series 1 is viable for timestamp 2 due to lookback). `info` picks series 2 since it has the newer sample (timestamp 2 versus timestamp 1).

The `info` function is disabled by default, and has to be enabled via `--enable-feature=promql-experimental-functions`.

The below screenshot demonstrates the `info` function, including three different data labels from `target_info` (`k8s_cluster_name`, `k8s_container_name` and `k8s_cronjob_name`).

![prometheus](https://github.com/user-attachments/assets/4099052f-bb89-4ce9-b6f3-83dc92254f97)

## Usage

Without the `info` function, to include `target_info` labels (Prometheus' encoding of OTel resource attributes), you have to write a join query like the following:

```promql
sum by (k8s_cluster_name, http_status_code) (
  rate(http_server_request_duration_seconds_count[2m])
  * on (job, instance) group_left (k8s_cluster_name)
  target_info
)
```

With the `info` function, you can instead of the above write:

```promql
sum by (k8s_cluster_name, http_status_code) (
  info(
    rate(http_server_request_duration_seconds_count[2m]),
    {k8s_cluster_name=~".+"}
  )
)
```

## Known problems

* The data label matcher auto completion is imprecise, in that it looks up labels for all time series instead of just relevant info series (by default: `target_info`)
  * Maybe this can easily be fixed in the frontend code? 

## Benchmark results

<details>
<summary>Benchmark stats for join query vs. <code>info</code> function</summary>

```console
✗ benchstat join.txt info.txt
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: AMD Ryzen 9 3950X 16-Core Processor            
                                                                  │   join.txt   │              info.txt              │
                                                                  │    sec/op    │   sec/op     vs base               │
InfoFunction/Joining_info_metrics_with_other_metrics_example_1-32    7.370m ± 6%   2.112m ± 7%  -71.35% (p=0.002 n=6)
InfoFunction/Joining_info_metrics_with_other_metrics_example_2-32   13.278m ± 3%   2.156m ± 3%  -83.76% (p=0.002 n=6)
geomean                                                              9.892m        2.134m       -78.43%

                                                                  │   join.txt   │              info.txt               │
                                                                  │     B/op     │     B/op      vs base               │
InfoFunction/Joining_info_metrics_with_other_metrics_example_1-32   452.3Ki ± 0%   397.2Ki ± 0%  -12.17% (p=0.002 n=6)
InfoFunction/Joining_info_metrics_with_other_metrics_example_2-32   705.7Ki ± 1%   400.6Ki ± 0%  -43.24% (p=0.002 n=6)
geomean                                                             564.9Ki        398.9Ki       -29.39%

                                                                  │  join.txt   │              info.txt              │
                                                                  │  allocs/op  │  allocs/op   vs base               │
InfoFunction/Joining_info_metrics_with_other_metrics_example_1-32   3.474k ± 0%   3.280k ± 0%   -5.60% (p=0.002 n=6)
InfoFunction/Joining_info_metrics_with_other_metrics_example_2-32   5.465k ± 0%   3.322k ± 0%  -39.21% (p=0.002 n=6)
geomean                                                             4.357k        3.301k       -24.24%
```
</details>